axiosConfigia vähän muutettu kun ei toiminut kirjautumissivu, AllShiftsissä toimii delete nappi, shiftServiceen lisätty pyyntöihin axiosConfig

+++ controllers/shifts.js
@@ -1,4 +1,4 @@
-import bcrypt from 'bcrypt'
+// import bcrypt from 'bcrypt'
 import express from 'express'
 // import jwt from 'jsonwebtoken'
 // import Company from '../models/company.js'
@@ -8,35 +8,37 @@ import Shift from '../models/shift.js'
 const shiftsRouter = express.Router();
 
 shiftsRouter.get('/', async (request, response) => {
-    const shifts = await Shift
-      .find({}).populate('employeeId', { id: 1, name: 1 })
-    response.json(shifts)
+  if(!request.user) {
+    return response.status(400).json({
+      error: 'must be logged in to see shifts'
+    })
+  }
+
+  const shifts = await Shift
+    .find({}).populate('employeeId', { id: 1, name: 1 })
+  response.json(shifts)
 })
 
 shiftsRouter.get('/:id', async (request, response) => {
+  if(!request.user) {
+    return response.status(400).json({
+      error: 'must be logged in to see messages'
+    })
+  }
+
   const shift = await Shift
     .findById(request.params.id)
   response.json(shift)
 })
 
 shiftsRouter.post('/', async (request, response) => {
+  if(!request.user) {
+    return response.status(400).json({
+      error: 'must be logged in to create new messages'
+    })
+  }
+
   const { date, start, end, employeeId, company } = request.body
-
-  // const existingUser = await User.findOne({ username })
-  // if (existingUser) {
-  //   return response.status(400).json({
-  //     error: 'username must be unique'
-  //   })
-  // }
-
-  // if (password.length < 3) {
-  //   return response.status(400).json({
-  //     error: 'password not entered or it is too short(minimum length 3)'
-  //   })
-  // }
-
-  // const saltRounds = 10
-  // const passwordHash = await bcrypt.hash(password, saltRounds)
 
   const shift = new Shift({
     date,
@@ -46,18 +48,25 @@ shiftsRouter.post('/', async (request, response) => {
     company,
   })
 
-  const savedShift = await shift.save()
-
-  const userToUpdate = await User.findById(employeeId)
-  
-  userToUpdate.shifts = userToUpdate.shifts.concat(savedShift._id)
-  
-  await userToUpdate.save()
-
+  try {
+    const savedShift = await shift.save()
+    const userToUpdate = await User.findById(employeeId)
+    
+    userToUpdate.shifts = userToUpdate.shifts.concat(savedShift._id)
+    
+    await userToUpdate.save()
+  }
+  catch(error) {
+  }
   response.status(201).json(savedShift)
 })
 
 shiftsRouter.put('/:id', async (request, response, next) => {
+  if(!request.user) {
+    return response.status(400).json({
+      error: 'must be logged in to edit messages'
+    })
+  }
   // const body = request.body
   // const token = getTokenFrom(request)
   // const decodedToken = jwt.verify(token, process.env.SECRET)
@@ -92,9 +101,14 @@ shiftsRouter.put('/:id', async (request, response, next) => {
 })
 
 shiftsRouter.delete('/:id', async (request, response) => {
-  //const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!request.user) {
+    return response.status(400).json({
+      error: 'must be logged in to delete messages'
+    })
+  }
+  
   const shift = await Shift.findById(request.params.id)
-  // console.log('user: ', user)
+  
   if (shift) {
     await Shift.findByIdAndDelete(request.params.id)
     response.status(204).end()
